Added cflags to Makefile and fixed the build errors

+++ src/main.c
@@ -15,7 +15,7 @@ int main(int argc, const char* argv[])
 
     if (parse_elf(argv[1]) == 0) {
         // instrc_arr and name_arr are extern from elfparser.h
-        for (int i = 0; i < instrc_arr.size; i++) {
+        for (size_t i = 0; i < instrc_arr.size; i++) {
             disassemble(instrc_arr.list[i].bits, instrc_arr.list[i].size, instrc_arr.list[i].offset, &name_arr);
         }
     }
